피 깍인 상태로 전투 시작하면 시작할때 피 쫙 죽어드는 애니메이션 고침

+++ Assets/01.Scripts/Unit/Unit.cs
@@ -17,7 +17,7 @@ public class Unit : MonoBehaviour
     public int MaxHP => _maxHealth;
 
     [SerializeField] private int _health = 10;
-    public int HP
+    public int HP 
     {
         get => _health;
         protected set
@@ -270,7 +270,7 @@ public class Unit : MonoBehaviour
 
     public void UISetting()
     {
-        _healthBar.DOScaleX(HP / MaxHP, 0);
+        _healthBar.DOScaleX((float)HP / MaxHP, 0);
         _healthFeedbackBar.DOScaleX(0, 0);
         _healthText.text = string.Format("{0}/{1}", HP.ToString(), MaxHP.ToString());
